Changed the minimum target operating system to XP so that the build would be compatible with the Windows 10 SDK.
Removed MoodWatch.h header file. MoodWatch will not be featured in the first cut of Hermes.

In the process of culling files that utilize PaigeDLL. I am pretty sure that Paige DLL was used for printing messages and maybe some other text manipulation. The first cut of Hermes will not be able to print hardcopy messages.